# performance - make build_tags_taxonomy() use string comparisons instead of regexps