@svensven 2020-09-12
this is what i was hoping to see :D
my other pull requests:
![image](https://user-images.githubusercontent.com/5736616/92998785-f6155b80-f513-11ea-943c-35eb630f31c0.png)
![image](https://user-images.githubusercontent.com/5736616/92998804-0299b400-f514-11ea-9c7b-dec3ca5e7c47.png)
this one:
![image](https://user-images.githubusercontent.com/5736616/92998812-0af1ef00-f514-11ea-92fd-7c3546b96d27.png)