@Ban3 2020-09-12
I guess there's no way avoiding the regex compilation because of the variables used? That's probably the cause for the slowness because those regexes never actually get to the regex engine but are optimized to substr searches:

```
Matching REx "-lovech-" against "sandanski"
UTF-8 pattern and string...
Intuit: trying to determine minimum start position...
  doing 'check' fbm scan, [0..9] gave -1
  Did not find anchored substr "-lovech-"...
Match rejected by optimizer
```