@stephanegigandet 2020-09-12
WOW, very cool results! That's going to make a lot of people very happy! :)

> It seems to produce identical .results.txt files (and the same complaints/errors to the console), but different but identical-size .sto files. I don't know if that's significant, maybe they're different every time?

From the Storable doc, it seems you can set $Storable::canonical = 1 to force the .sto file to have sorted keys, that should allow to verify that the resulting structures are exactly the same:

"Normally Storable stores elements of hashes in the order they are stored internally by Perl, i.e. pseudo-randomly. If you set $Storable::canonical to some TRUE value, Storable will store hashes with the elements sorted by their key. This allows you to compare data structures by comparing their frozen representations (or even the compressed frozen representations), which can be useful for creating lookup tables for complicated queries."

https://metacpan.org/pod/Storable#CANONICAL-REPRESENTATION