@stephanegigandet 2020-09-12
@hangy also used that rindex trick to speed up the Store::get_string_id_for_lang function that we use billions of times (during the taxonomy building but also very extensively when we analyze ingredients, serve queries etc.).