@svensven 2020-09-12
I put ```$Storable::canonical = 1;``` in build_tags_taxonomy.pl and the .sto files came out identical in both versions.